@@ -12,14 +12,15 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^7.2.4",
-    "@angular/cdk": "^7.3.1",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/cdk": "^7.3.2",
+    "@angular/common": "~7.2.4",
+    "@angular/compiler": "~7.2.4",
+    "@angular/core": "~7.2.4",
+    "@angular/forms": "~7.2.4",
+    "@angular/material": "^7.3.2",
+    "@angular/platform-browser": "~7.2.4",
+    "@angular/platform-browser-dynamic": "~7.2.4",
+    "@angular/router": "~7.2.4",
     "@ng-bootstrap/ng-bootstrap": "^4.0.1",
     "@ngx-gallery/core": "^4.0.2-beta.0",
     "@ngx-gallery/gallerize": "^4.0.2-beta.0",
@@ -28,20 +29,21 @@
     "@ngx-translate/http-loader": "^4.0.0",
     "bootstrap": "^4.2.1",
     "core-js": "^2.5.4",
+    "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "ngx-lightbox": "^1.2.0",
     "ngx-pagination": "^3.2.1",
     "popper.js": "^1.14.6",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.4.0",
     "rxjs-compat": "^6.4.0",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.8.29"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.0",
-    "@angular/cli": "~7.1.2",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular/cli": "^7.3.1",
+    "@angular/compiler-cli": "~7.2.4",
+    "@angular/language-service": "~7.2.4",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -56,6 +58,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.2.4"
   }
 }
