@@ -1,3 +1,9 @@
+import {
+  MatButtonModule,
+  MatCheckboxModule,
+  MatFormFieldModule,
+  MatProgressSpinnerModule
+} from '@angular/material';
 import { NgModule } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { GalleryModule } from '@ngx-gallery/core';
@@ -8,11 +14,13 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { RestaurantCardComponent } from './pages/home/restaurant-card/restaurant-card.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
 import { FooterComponent } from './components/footer/footer.component';
 import { HeaderComponent } from './components/header/header.component';
 import { ContactComponent } from './pages/contact/contact.component';
+import {
+  RestaurantCardComponent
+} from './pages/home/restaurant-card/restaurant-card.component';
 import { AboutComponent } from './pages/about/about.component';
 import { HomeComponent } from './pages/home/home.component';
 import { AppComponent } from './components/app.component';
@@ -56,7 +64,11 @@ export function HttpLoaderFactory(http: HttpClient) {
       loadingMode: 'indeterminate',
     }),
     LightboxModule,
-    GallerizeModule
+    GallerizeModule,
+    MatProgressSpinnerModule,
+    MatButtonModule,
+    MatCheckboxModule,
+    MatFormFieldModule,
   ],
   providers: [
     /*{
