@@ -1,7 +1,7 @@
-import {Observable} from 'rxjs';
-import {Injectable} from '@angular/core';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/do';
+import { Observable } from 'rxjs';
+import { Injectable } from '@angular/core';
 import {
   HttpEvent,
   HttpInterceptor,
@@ -44,6 +44,11 @@ export class TokenInterceptor implements HttpInterceptor {
             }, (error) => {
               console.log(error);
             });
+        }
+        if (err.error.message === 'RefreshToken expired!') {
+          console.log(err.error.message);
+          StorageService.clearItem('accessToken');
+          StorageService.clearItem('refreshToken');
         }
       }
     });
