@@ -1,5 +1,6 @@
 export * from './restaurants.model';
 export * from './products.model';
-export * from './orderProducts.model';
-export * from './users.model';
+export * from './order.model';
 export * from './gallery.model';
+export * from './users.model';
+export * from './time.model';
