@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { TranslateModule } from '@ngx-translate/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { OrderedProductsComponent } from './ordered-products/ordered-products.component';
-import { OrderRequestService } from '../../core/services/order-request.service';
+import { OrderHelperService } from '../../core/services/orderHelper.service';
 import { OrderModalComponent } from './order-modal/order-modal.component';
 import { ValidateService } from '../../core/services/validate.service';
+import { OrderService } from '../../core/services/order.service';
 import { SharedModule } from '../../shared/shared.module';
 import { OrderComponent } from './order.component';
 
@@ -19,14 +21,16 @@ import { OrderComponent } from './order.component';
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
-    SharedModule
+    SharedModule,
+    TranslateModule
   ],
   exports: [
     OrderModalComponent,
     OrderComponent
   ],
   providers: [
-    OrderRequestService,
+    OrderHelperService,
+    OrderService,
     ValidateService
   ]
 })
