@@ -1,20 +1,39 @@
 import { Component, OnInit } from '@angular/core';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import { ActivatedRoute } from '@angular/router';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { NgbDatepickerConfig, NgbTimepickerConfig } from '@ng-bootstrap/ng-bootstrap';
+
+import { ReservationService } from '../../../../core/services/reservation.service';
+import { ValidateService } from '../../../../core/services/validate.service';
+import { TimeModel } from '../../../../core/models';
+
+declare const $: any;
 
 @Component({
   selector: 'app-make-reservation-form',
   templateUrl: './make-reservation-form.component.html',
-  styleUrls: ['./make-reservation-form.component.scss']
+  styleUrls: ['./make-reservation-form.component.scss'],
+  providers: [NgbTimepickerConfig]
 })
 export class MakeReservationFormComponent implements OnInit {
 
   public reserveForm: FormGroup;
-  public timer = {hour: 13, minute: 30};
+  public startTime: TimeModel = { hour: 13, minute: 30 };
+  public endTime: TimeModel = { hour: 15, minute: 10 };
   public date: FormControl;
-  // public time: FormControl;
-  public for: FormControl;
+  public guestsCount: FormControl;
+  public showSpinner = false;
+  public tables: Array<any>;
 
-  constructor() {
+  constructor(private config: NgbTimepickerConfig,
+              private dateConfig: NgbDatepickerConfig,
+              private reservationService: ReservationService,
+              private route: ActivatedRoute) {
+
+    const currentDate = new Date();
+    config.spinners = false;
+    dateConfig.minDate = {year: currentDate.getFullYear(), month: currentDate.getMonth() + 1, day: currentDate.getDate()};
+    dateConfig.outsideDays = 'hidden';
   }
 
   ngOnInit() {
@@ -24,46 +43,56 @@ export class MakeReservationFormComponent implements OnInit {
 
   private createFormControls(): void {
     this.date = new FormControl('', [Validators.required]);
-    this.for = new FormControl('1');
+    this.guestsCount = new FormControl('1');
   }
 
   private createForm(): void {
     this.reserveForm = new FormGroup({
       'date': this.date,
-      'for': this.for,
-    });
-  }
-
-  public validateAllFormFields(formGroup: FormGroup): void {
-    Object.keys(formGroup.controls).forEach(field => {
-      const control = formGroup.get(field);
-
-      if (control instanceof FormControl) {
-        control.markAsTouched({onlySelf: true});
-        console.log('ok');
-      } else if (control instanceof FormGroup) {
-        this.validateAllFormFields(control);
-      }
+      'guestsCount': this.guestsCount,
     });
   }
 
   public alertValidate(event) {
-    if (this[event.name].status === 'INVALID') {
-      event.classList.add('bg-danger');
-    } else {
-      event.classList.remove('bg-danger');
-    }
+    ValidateService.alertValidate(event, this.reserveForm);
   }
 
-  submitReservForm() {
-    console.log((this.date.dirty && this.date.hasError('required'))
-      || (this.date.touched && this.date.hasError('required')));
+  private checkTimeCharacter(time) {
+    time.hour = time.hour.toString().length === 2 ? time.hour : '0' + time.hour;
+    time.minute = time.minute.toString().length === 2 ? time.minute : '0' + time.minute;
+  }
+
+  submitReserveForm() {
     if (this.reserveForm.invalid) {
-      this.validateAllFormFields(this.reserveForm);
-    } else {
-      console.log(this.reserveForm.getRawValue());
+      return ValidateService.validateAllFormFields(this.reserveForm);
     }
+    this.showSpinner = true;
+    const {date, guestsCount} = this.reserveForm.getRawValue();
+    this.checkTimeCharacter(this.startTime);
+    this.checkTimeCharacter(this.endTime);
+    ReservationService.request = {
+      date: `${date.year}-${date.month}-${date.day}`,
+      guestsCount,
+      startTime: this.startTime,
+      endTime: this.endTime,
+      restaurantId: this.route.snapshot.params.id,
+    };
+    this.reservationService.checkFreeTable(ReservationService.request)
+      .subscribe((data) => {
+        if (!data.data.length) {
+          this.tables = null;
+          this.showSpinner = false;
+          return $('#reservation').modal('show');
+        }
+        data.data.forEach((time) => {
+          this.checkTimeCharacter(time.startTime);
+          this.checkTimeCharacter(time.endTime);
+        });
+        this.tables = data.data;
+      }, (error) => {
+        this.showSpinner = false;
+        console.log(error);
+      });
   }
-
 
 }
